changed mag to int32_t to match isqrt function and prevent overflow

+++ examples/EXAMPLE01_Basic/EXAMPLE01_Basic.ino
@@ -2,7 +2,7 @@
  * FILENAME: EXAMPLE01_basic.ino
  * AUTHOR:   Orlando S. Hoilett
  * CONTACT:  orlandohoilett@gmail.com
- * VERSION:  2.0.0
+ * VERSION:  3.0.0
  * 
  * 
  * AFFILIATIONS
@@ -29,6 +29,10 @@
  * Version 2.0.0
  * 2020/08/21:1847>
  *           - Moved to a templated class.
+ * Version 3.0.0
+ * 2020/08/23:0406> (UTC-5)
+ *           - changed magnitude types to int32_t to
+ *           match isqrt function and prevent overflow
  * 
  * 
  * DISCLAIMER
@@ -77,7 +81,7 @@ void setup()
   Serial.begin(9600); //Use SerialUSB for SparkFun SAMD21 boards
   while(!Serial); //will not run until Serial Monitor is open
 
-  int16_t mag[samples] = {0};
+  int32_t mag[samples] = {0};
   uint16_t startIndex = 0;
   uint16_t endIndex = 0;
   KickFFT<int16_t>::fft(Fs, 0, Fs/2, samples, input, mag, startIndex, endIndex);
